better support for more field types in timestampdiff()/timestampadd()
- millisecond/microsecond/week/quarter now work in timestampadd()
- had to change the return type of typestampdiff() to Long because
  years->milliseconds don't fit in Integer
- also improve handling of millisecond/microsecond in Postgres for
  timestampdiff()
- improved generated SQL for timestampadd() on Postgres

+++ hibernate-core/src/main/java/org/hibernate/dialect/DB2Dialect.java
@@ -183,7 +183,7 @@ public class DB2Dialect extends Dialect {
 				.setExactArgumentCount( 3 )
 				.register();
 		queryEngine.getSqmFunctionRegistry().patternTemplateBuilder( "timestampdiff", "?1s_between(?3,?2)" )
-				.setInvariantType( StandardSpiBasicTypes.INTEGER )
+				.setInvariantType( StandardSpiBasicTypes.LONG )
 				.setExactArgumentCount( 3 )
 				.register();
